REFACT: move rbloader and speed tools in sub package
